fix(meal-recognition-backend) split all models and types to seperate files
