+++ meal-analysis-service/src/meal-recognition/models/ingredient-details.interface.ts
@@ -1,0 +1,8 @@
+import { Per100gNutrition } from "../types/nutrition.types";
+
+export interface IngredientDetails {
+    name: string;
+    weight: number;
+    usdaFoodLabel?: string;
+    per100gNutrition?: Per100gNutrition;
+}
