+++ meal-analysis-service/src/meal-recognition/models/nutrition-details.interface.ts
@@ -1,0 +1,4 @@
+export interface NutritionDetails {
+    value: number;
+    unit: string;
+}
