--- meal-analysis-service/src/meal-recognition/types/ingredient-details.interface.ts
@@ -1,25 +1,0 @@
-export interface IngredientDetails {
-  name: string;
-  weight: number;
-  usdaFoodLabel?: string;
-  per100gNutrition?: Per100gNutrition;
-}
-
-export type Per100gNutrition = Record<Nutrient | string, NutritionDTO>;
-
-export interface NutritionDTO {
-  value: number;
-  unit: string;
-}
-
-export type Nutrient =
-  | 'calories'
-  | 'protein'
-  | 'carbohydrates'
-  | 'fat'
-  | 'iron'
-  | 'fiber'
-  | 'vitaminC'
-  | 'vitaminA'
-  | 'calcium'
-  | 'magnesium';
