+++ meal-analysis-service/src/meal-recognition/types/nutrition.types.ts
@@ -1,0 +1,15 @@
+import { NutritionDetails } from "../models/nutrition-details.interface";
+
+export type Per100gNutrition = Record<Nutrient | string, NutritionDetails>;
+
+export type Nutrient =
+  | 'calories'
+  | 'protein'
+  | 'carbohydrates'
+  | 'fat'
+  | 'iron'
+  | 'fiber'
+  | 'vitaminC'
+  | 'vitaminA'
+  | 'calcium'
+  | 'magnesium';
